requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
ultralytics>=8.0.0
segmentation-models-pytorch>=0.3.0
ray[tune]>=2.8.0
mlflow>=2.8.0

# Computer Vision
timm>=0.9.0
transformers>=4.35.0
albumentations>=1.3.0
pycocotools>=2.0.0
opencv-python>=4.8.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Databricks integration
databricks-cli>=0.17.0
databricks-sdk>=0.12.0

# Testing and development
pytest>=7.4.0
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
torchmetrics>=1.0.0 